# docs/swagger.yaml
definitions:
  errcode.Error:
    properties:
      code:
        type: integer
      details:
        items:
          type: string
        type: array
      err:
        type: error
      msg:
        type: string
    type: object
...
  service.CategoryResponse:
    properties:
      id:
        example: 1
        type: integer
      name:
        example: 电影
        type: string
      workNum:
        example: 1
        type: integer
    type: object
...
  service.NewsPageResponse:
    properties:
      list:
        items:
          properties:
            commentNums:
              type: integer
            description:
              type: string
            id:
              type: integer
            title:
              type: string
            updatedAt:
              type: string
          type: object
        type: array
      total:
        type: integer
    type: object
  service.NewsRequest:
    properties:
      content:
        type: string
      title:
        type: string
    required:
    - content
    - title
    type: object
  service.NewsResponse:
    properties:
      commentNums:
        type: integer
      content:
        type: string
      id:
        type: integer
      title:
        type: string
      updatedAt:
        type: string
    type: object
  service.NewsUpdateRequest:
    properties:
      content:
        type: string
      title:
        type: string
    type: object
  service.ReplyRequest:
    properties:
      commentId:
        type: integer
      content:
        type: string
    required:
    - commentId
    - content
    type: object
  service.ReplyResponse:
    properties:
      content:
        type: string
      createdAt:
        type: string
      id:
        type: integer
      user:
        $ref: '#/definitions/service.UserResponse'
        type: object
    type: object
  service.SubscriptionRequest:
    properties:
      objId:
        type: integer
      objType:
        type: integer
    required:
    - objId
    - objType
    type: object
  service.SubsectionRequest:
    properties:
      introduce:
        type: string
      name:
        type: string
    type: object
  service.SubsectionResponse:
    properties:
      id:
        type: integer
      introduce:
        type: string
      name:
        type: string
      seq:
        type: integer
      workNum:
        type: integer
    type: object
...
  service.TopicPageResponse:
    properties:
      list:
        items:
          $ref: '#/definitions/service.TopicResponse'
        type: array
      total:
        type: integer
    type: object
  service.TopicRequest:
    properties:
      categoryId:
        type: integer
      description:
        type: string
      original:
        type: string
      originalUrl:
        type: string
      title:
        type: string
    required:
    - categoryId
    - description
    - original
    - originalUrl
    - title
    type: object
  service.TopicResponse:
    properties:
      category:
        $ref: '#/definitions/service.CategoryResponse'
        type: object
      createdAt:
        type: string
      description:
        type: string
      id:
        type: integer
      original:
        type: string
      originalUrl:
        type: string
      subscribeNums:
        type: integer
      title:
        type: string
      updatedAt:
        type: string
      workNum:
        type: integer
    type: object
  service.TopicUpdateRequest:
    properties:
      categoryId:
        type: integer
      description:
        type: string
      original:
        type: string
      originalUrl:
        type: string
      title:
        type: string
    type: object
...
  service.UserResponse:
    properties:
      avatar:
        type: string
      createdAt:
        type: string
      email:
        type: string
      fansNums:
        type: integer
      gender:
        type: integer
      id:
        type: integer
      introduce:
        type: string
      root:
        type: boolean
      username:
        type: string
      words:
        type: integer
      workDay:
        type: integer
      worksNums:
        type: integer
    type: object
...
info:
  contact: {}
  description: 属于国内的同人圈
  license: {}
  termsOfService: https://github.com/shyptr/archiveofourown
  title: 同人圈
  version: "1.0"
paths:
...
  /api/v1/news:
    post:
      consumes:
      - application/json
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 请求参数
        in: body
        name: param
        required: true
        schema:
          $ref: '#/definitions/service.NewsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            type: string
        "400":
          description: 请求错误
          schema:
            $ref: '#/definitions/errcode.Error'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/errcode.Error'
      summary: 添加公告
      tags:
      - 公告
  /api/v1/news/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 公告ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            type: string
        "400":
          description: 请求错误
          schema:
            $ref: '#/definitions/errcode.Error'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/errcode.Error'
      summary: 删除公告
      tags:
      - 公告
    get:
      consumes:
      - application/json
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 公告ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/service.NewsResponse'
        "400":
          description: 请求错误
          schema:
            $ref: '#/definitions/errcode.Error'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/errcode.Error'
      summary: 公告内容
      tags:
      - 公告
    put:
      consumes:
      - application/json
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 公告ID
        in: path
        name: id
        required: true
        type: integer
      - description: 请求参数
        in: body
        name: param
        required: true
        schema:
          $ref: '#/definitions/service.NewsUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            type: string
        "400":
          description: 请求错误
          schema:
            $ref: '#/definitions/errcode.Error'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/errcode.Error'
      summary: 修改公告
      tags:
      - 公告
  /api/v1/news/list:
    get:
      consumes:
      - application/json
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 页大小
        in: query
        name: pageSize
        required: true
        type: integer
      - description: 当前页
        in: query
        name: page
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/service.NewsPageResponse'
        "400":
          description: 请求错误
          schema:
            $ref: '#/definitions/errcode.Error'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/errcode.Error'
      summary: 公告列表
      tags:
      - 公告
...
  /api/v1/reply:
    post:
      consumes:
      - application/json
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 请求参数
        in: body
        name: param
        required: true
        schema:
          $ref: '#/definitions/service.ReplyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            type: string
        "400":
          description: 请求错误
          schema:
            $ref: '#/definitions/errcode.Error'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/errcode.Error'
      summary: 添加回复
      tags:
      - 回复
  /api/v1/reply/{commentID}:
    get:
      consumes:
      - application/json
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 评论ID
        in: path
        name: commentID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            items:
              $ref: '#/definitions/service.ReplyResponse'
            type: array
        "400":
          description: 请求错误
          schema:
            $ref: '#/definitions/errcode.Error'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/errcode.Error'
      summary: 回复列表
      tags:
      - 回复
  /api/v1/reply/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 回复ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            type: string
        "400":
          description: 请求错误
          schema:
            $ref: '#/definitions/errcode.Error'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/errcode.Error'
      summary: 回复列表
      tags:
      - 回复
  /api/v1/subscription:
    delete:
      consumes:
      - application/json
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 请求参数
        in: body
        name: param
        required: true
        schema:
          $ref: '#/definitions/service.SubscriptionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            type: string
        "400":
          description: 请求错误
          schema:
            $ref: '#/definitions/errcode.Error'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/errcode.Error'
      summary: 添加订阅
      tags:
      - 订阅
    post:
      consumes:
      - application/json
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 请求参数
        in: body
        name: param
        required: true
        schema:
          $ref: '#/definitions/service.SubscriptionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            type: string
        "400":
          description: 请求错误
          schema:
            $ref: '#/definitions/errcode.Error'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/errcode.Error'
      summary: 添加订阅
      tags:
      - 订阅
  /api/v1/subsection:
    post:
      consumes:
      - application/json
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 请求参数
        in: body
        name: param
        required: true
        schema:
          $ref: '#/definitions/service.SubsectionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            type: string
        "400":
          description: 请求错误
          schema:
            $ref: '#/definitions/errcode.Error'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/errcode.Error'
      summary: 分卷列表
      tags:
      - 分卷
  /api/v1/subsection/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 分卷ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            type: string
        "400":
          description: 请求错误
          schema:
            $ref: '#/definitions/errcode.Error'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/errcode.Error'
      summary: 分卷列表
      tags:
      - 分卷
    put:
      consumes:
      - application/json
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 分卷ID
        in: path
        name: id
        required: true
        type: integer
      - description: 请求参数
        in: body
        name: param
        required: true
        schema:
          $ref: '#/definitions/service.SubsectionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            type: string
        "400":
          description: 请求错误
          schema:
            $ref: '#/definitions/errcode.Error'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/errcode.Error'
      summary: 分卷列表
      tags:
      - 分卷
  /api/v1/subsection/{workID}:
    get:
      consumes:
      - application/json
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 作品ID
        in: path
        name: workID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            items:
              $ref: '#/definitions/service.SubsectionResponse'
            type: array
        "400":
          description: 请求错误
          schema:
            $ref: '#/definitions/errcode.Error'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/errcode.Error'
      summary: 分卷列表
      tags:
      - 分卷
  /api/v1/tag/{name}:
    get:
      consumes:
      - application/json
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 标签名
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            items:
              type: string
            type: array
        "400":
          description: 请求错误
          schema:
            $ref: '#/definitions/errcode.Error'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/errcode.Error'
      summary: 查询匹配标签
      tags:
      - 标签
  /api/v1/topic:
    get:
      consumes:
      - application/json
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 页大小
        in: query
        name: pageSize
        required: true
        type: integer
      - description: 当前页
        in: query
        name: page
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/service.TopicPageResponse'
        "400":
          description: 请求错误
          schema:
            $ref: '#/definitions/errcode.Error'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/errcode.Error'
      summary: 专题列表
      tags:
      - 专题
    post:
      consumes:
      - application/json
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 请求参数
        in: body
        name: param
        required: true
        schema:
          $ref: '#/definitions/service.TopicRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            type: string
        "400":
          description: 请求错误
          schema:
            $ref: '#/definitions/errcode.Error'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/errcode.Error'
      summary: 创建专题
      tags:
      - 专题
  /api/v1/topic/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 专题ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            type: string
        "400":
          description: 请求错误
          schema:
            $ref: '#/definitions/errcode.Error'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/errcode.Error'
      summary: 修改专题
      tags:
      - 专题
    put:
      consumes:
      - application/json
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 专题ID
        in: path
        name: id
        required: true
        type: integer
      - description: 请求参数
        in: body
        name: param
        required: true
        schema:
          $ref: '#/definitions/service.TopicUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            type: string
        "400":
          description: 请求错误
          schema:
            $ref: '#/definitions/errcode.Error'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/errcode.Error'
      summary: 修改专题
      tags:
      - 专题
...
swagger: "2.0"